requirements.txt
# Core AI/ML Dependencies
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.43
langchain-google-genai==1.0.5
langchain-qdrant==0.2.1
qdrant-client==1.15.0


# Re-ranking
sentence-transformers==5.1.0

# Code Processing
gitpython==3.1.43
tiktoken==0.6.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
esprima==4.0.1

# Search & Retrieval
rank-bm25==0.2.2
simsimd==3.7.7

# Web Framework
fastapi==0.124.4
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.12.0
requests==2.32.5
httpx==0.28.1
aiofiles==24.1.0

# Database
sqlalchemy==2.0.46
alembic==1.16.4
psycopg2-binary==2.9.11
asyncpg==0.31.0
aiosqlite==0.21.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Resilience
tenacity==8.5.0

# Configuration
python-dotenv==1.2.1

# Development & Quality Tools
pytest==8.4.1
pytest-asyncio==1.1.0
black==26.1.0
mypy==1.19.1
ruff==0.13.1
locust==2.23.1
# Observability
sentry-sdk==1.45.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0

# Caching & Background Jobs
cachetools==5.3.2
celery==5.3.6
redis==5.0.3

